Update: move function.
