+++ server/src/services/definition.ts
@@ -1,17 +1,11 @@
 import {
   DefinitionLink,
-  LocationLink,
   Logger,
   Position,
-  URI,
 } from "vscode-languageserver"
 import { TextDocument } from "vscode-languageserver-textdocument"
 
-import { parseCreateStatements } from "@/postgres/parsers/parseCreateStatements"
-import { parseStmtements } from "@/postgres/parsers/statement"
-import {
-  Definition, DefinitionName, DefinitionsManager,
-} from "@/server/definitionsManager"
+import { DefinitionsManager } from "@/server/definitionsManager"
 import { sanitizeWordCandidates } from "@/utilities/sanitizeWord"
 import { getWordRangeAtPosition } from "@/utilities/text"
 
@@ -40,57 +34,4 @@ export async function getDefinitionLinks(
   }
 
   return undefined
-}
-
-
-export async function parseDefinitions(
-  uri: URI,
-  fileText: string,
-  defaultSchema: string,
-  logger: Logger,
-): Promise<Definition[] | undefined> {
-  const statements = await parseStmtements(uri, fileText, logger)
-  if (statements === undefined) {
-    return undefined
-  }
-
-
-  return parseCreateStatements(fileText, statements).flatMap(
-    (statementInfo) => {
-      return makeMultiSchemaDefinitions(
-        statementInfo.name,
-        LocationLink.create(
-          uri,
-          statementInfo.targetRange,
-          statementInfo.targetSelectionRange,
-        ),
-        statementInfo.schema,
-        defaultSchema,
-      )
-    },
-  )
-}
-
-function makeMultiSchemaDefinitions(
-  name: DefinitionName,
-  link: DefinitionLink,
-  schema: string | undefined,
-  defaultSchema: string,
-): Definition[] {
-  const definitions = [
-    {
-      name: (schema ?? defaultSchema) + "." + name,
-      link,
-    },
-  ]
-
-  // On the default schema, add definition without schema.
-  if (schema === undefined || schema === defaultSchema) {
-    definitions.push({
-      name,
-      link,
-    })
-  }
-
-  return definitions
 }
